+++ tests/ext/aiohttp/test_middleware.py
@@ -7,6 +7,7 @@ import asyncio
 from unittest.mock import patch
 
 from aiohttp import web
+from aiohttp.web_exceptions import HTTPUnauthorized
 import pytest
 
 from aws_xray_sdk.core.emitters.udp_emitter import UDPEmitter
@@ -48,13 +49,24 @@ class ServerTest(object):
         """
         Handle / request
         """
-        return web.Response(text="ok")
+        if "content_length" in request.query:
+            headers = {'Content-Length': request.query['content_length']}
+        else:
+            headers = None
+
+        return web.Response(text="ok", headers=headers)
 
     async def handle_error(self, request: web.Request) -> web.Response:
         """
         Handle /error which returns a 404
         """
         return web.Response(text="not found", status=404)
+
+    async def handle_unauthorized(self, request: web.Request) -> web.Response:
+        """
+        Handle /unauthorized which returns a 401
+        """
+        raise HTTPUnauthorized()
 
     async def handle_exception(self, request: web.Request) -> web.Response:
         """
@@ -74,6 +86,7 @@ class ServerTest(object):
         app.router.add_get('/', self.handle_ok)
         app.router.add_get('/error', self.handle_error)
         app.router.add_get('/exception', self.handle_exception)
+        app.router.add_get('/unauthorized', self.handle_unauthorized)
         app.router.add_get('/delay', self.handle_delay)
 
         return app
@@ -122,6 +135,41 @@ async def test_ok(test_client, loop, recorder):
     assert request['method'] == 'GET'
     assert request['url'] == 'http://127.0.0.1:{port}/'.format(port=client.port)
     assert response['status'] == 200
+
+
+async def test_ok_x_forwarded_for(test_client, loop, recorder):
+    """
+    Test a normal response with x_forwarded_for headers
+
+    :param test_client: AioHttp test client fixture
+    :param loop: Eventloop fixture
+    :param recorder: X-Ray recorder fixture
+    """
+    client = await test_client(ServerTest.app(loop=loop))
+
+    resp = await client.get('/', headers={'X-Forwarded-For': 'foo'})
+    assert resp.status == 200
+
+    segment = recorder.emitter.pop()
+    assert segment.http['request']['client_ip'] == 'foo'
+    assert segment.http['request']['x_forwarded_for']
+
+
+async def test_ok_content_length(test_client, loop, recorder):
+    """
+    Test a normal response with content length as response header
+
+    :param test_client: AioHttp test client fixture
+    :param loop: Eventloop fixture
+    :param recorder: X-Ray recorder fixture
+    """
+    client = await test_client(ServerTest.app(loop=loop))
+
+    resp = await client.get('/?content_length=100')
+    assert resp.status == 200
+
+    segment = recorder.emitter.pop()
+    assert segment.http['response']['content_length'] == 100
 
 
 async def test_error(test_client, loop, recorder):
@@ -174,6 +222,31 @@ async def test_exception(test_client, loop, recorder):
     assert request['client_ip'] == '127.0.0.1'
     assert response['status'] == 500
     assert exception.type == 'KeyError'
+
+
+async def test_unhauthorized(test_client, loop, recorder):
+    """
+    Test a 401 response
+
+    :param test_client: AioHttp test client fixture
+    :param loop: Eventloop fixture
+    :param recorder: X-Ray recorder fixture
+    """
+    client = await test_client(ServerTest.app(loop=loop))
+
+    resp = await client.get('/unauthorized')
+    assert resp.status == 401
+
+    segment = recorder.emitter.pop()
+    assert not segment.in_progress
+    assert segment.error
+
+    request = segment.http['request']
+    response = segment.http['response']
+    assert request['method'] == 'GET'
+    assert request['url'] == 'http://127.0.0.1:{port}/unauthorized'.format(port=client.port)
+    assert request['client_ip'] == '127.0.0.1'
+    assert response['status'] == 401
 
 
 async def test_response_trace_header(test_client, loop, recorder):